fixing logic
Previous policy would prevent the creation of additional ip configurations on a NIC even when it wasn't part of a backendpool as the backendpool element would exist during the update of the ipconfig. I'm now testing to see if there actually is a config in that element and not just the existence of that element.