Create azurepolicy.json
Policy to deploy a default rule to a NSG. 
Platform teams would like to enforce certain rules in a NSG. This policy shows an example with "Deny RDP". Along with this there should be another policy which disallows a NSG rule with a priority of 100 (or a range) which is reserved for the cloud platform team.