Bugfix edge-case
a bugfix to address an edge-case that was discovered - supplying an empty array in the arm/bicep template for dns servers caused the first "anyOf" check to pass, despite it being empty, so the policy failed to apply as intended.  The fix forces the test to be against an array that has something (anything!) in it, so that the test then moves on to the second part to test against the allowed list.